requirements.txt
prometheus-client
pyserial